requirements.txt
Bottleneck==1.3.4
nltk==3.7
nose==1.3.7
notebook==6.4.11
numba==0.55.1
numexpr==2.8.1
numpy==1.22.3
numpydoc==1.3.1
opencv-python==4.5.5.64
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pkginfo==1.8.2
pluggy==1.0.0
ply==3.11
PyYAML==6.0
pyzmq==22.3.0
SQLAlchemy==1.4.36
statsmodels==0.13.2
streamlit==1.8.1
sympy==1.10.1
tqdm==4.64.0
